Added fit with pressure to forecast module
Created forecastMarder with two user routines. The first of them takes in daily well production data together with pressure data and fits the production data to the pressure-adjusted model.

The second routine plots a comparison of the model output on top of the original production data.

Usage is illustrated with the Jupyter file
FitWellWithPressureData.ipynb